添加项目新增价格参数

+++ common/src/main/java/com/ligg/common/entity/ProjectEntity.java
@@ -18,6 +18,6 @@ public class ProjectEntity {
     @JsonFormat(shape = JsonFormat.Shape.STRING)
     private Long projectId;
     private String projectName;
-    private Double projectPrice;// 项目价格
+    private Float projectPrice;// 项目价格
     private LocalDateTime projectCreatedAt;// 创建时间
 }
